% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descStatTable.R
\name{buildInference}
\alias{buildInference}
\title{Build inference table}
\usage{
buildInference(infStat, pCol = "p", id_row = FALSE)
}
\arguments{
\item{infStat}{Data.frame of beta, lower, upper, and p. The column names must include BETA, LOWER, UPPER and P)}
}
\value{
data.frame
}
\description{
Builds the inference table
}
\details{
This function builds the inference table for beta, upper, and lower
}
\author{
Jochen Kruppa
}
