% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Calibration_tbl}
\alias{Calibration_tbl}
\title{Calibration example table}
\format{
A data frame with one row per calibration observation.
}
\source{
Example data included with MSlineaR.
}
\usage{
Calibration_tbl
}
\description{
Calibration data used in the targeted MSlineaR example.
}
\keyword{datasets}
