% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{dir0}
\alias{dir0}
\title{Small wrapper function for dir}
\usage{
dir0(...)
}
\arguments{
\item{...}{}
}
\value{
file.path
}
\description{
Small wrapper function for dir
}
\details{
Small wrapper function for dir
}
\author{
Jochen Kruppa
}
