% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Feature_tbl_wide}
\alias{Feature_tbl_wide}
\title{Feature data in wide format}
\format{
A data frame with sample information in rows and feature signal
intensities in separate columns.
}
\source{
Example data included with MSlineaR.
}
\usage{
Feature_tbl_wide
}
\description{
Example LC-MS feature data in wide format.
}
\keyword{datasets}
