% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierDetection.R
\name{outlierDetection}
\alias{outlierDetection}
\title{Title}
\usage{
outlierDetection(
  dat,
  y = Intensity,
  x = Concentration,
  model = c("logistic", "linear", "quadratic"),
  res = 2,
  threshCor = 0.99,
  numboutlier = 1
)
}
\arguments{
\item{numboutlier}{}
}
\description{
Title
}
