% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plotFDS}
\alias{plotFDS}
\title{Title}
\usage{
plotFDS(
  LR_object,
  printPDF = TRUE,
  groupIndices = "all",
  statusLinear = c(TRUE, FALSE),
  ID = "all",
  outputfileName = c("Calibrationplot"),
  pdfwidth = 9,
  pdfheight = 60
)
}
\arguments{
\item{pdfheight}{}
}
\description{
Title
}
