% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plotSamples}
\alias{plotSamples}
\title{Title}
\usage{
plotSamples(
  LR_object,
  data_Sample,
  groupIndices = "all",
  statusLinear = c(TRUE, FALSE),
  ID = "all",
  printPDF = TRUE,
  outputfileName = c("Sampleplot"),
  pdfwidth = 9,
  pdfheight = 60
)
}
\arguments{
\item{pdfheight}{}
}
\description{
Title
}
