% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignals.R
\name{plotSignals}
\alias{plotSignals}
\title{Title}
\usage{
plotSignals(
  dat,
  x = "Concentration",
  y = "Intensity",
  Replicate = "Batch",
  nrcol = 2
)
}
\arguments{
\item{y}{}
}
\description{
Title
}
