% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descStatTable.R
\name{prepareDeskStatTable}
\alias{prepareDeskStatTable}
\title{Descriptive and inference statistics for a given data set}
\usage{
prepareDeskStatTable(
  data,
  infStat = NULL,
  continous_flag = NULL,
  cate_threshold = 12,
  grouped.by = NULL,
  ratio.by.col = TRUE,
  caption = "This is the caption",
  colnames = NULL,
  color_row = FALSE
)
}
\arguments{
\item{data}{Data set of descriptive variables. No response!}

\item{infStat}{Inference data set with the column names, including: id, beta, lower, upper, p}

\item{continous_flag}{Which of the variables should be continous?
If not given, automatically decision by cate_threshold}

\item{cate_threshold}{Cutoff of classes when a variable should be considered has continous?}

\item{grouped.by}{Should the descriptive statistics be grouped by this factor?}

\item{ratio.by.col}{Should be the ratio of the grouped variable calculated by column?}

\item{caption}{Caption of the generated pandoc table}

\item{colnames}{The column names of the generated pandoc table}
}
\value{
pandoc.table
}
\description{
Generate output table for pandoc
}
\details{
This function generates a descriptive table and adds the inference
statistics (must be given). Due to many changes this function is really big.
}
\examples{
data(iris)
irisInfStat <- data.frame(id = names(iris)[-5],
                          beta = c(1, 2, 3, 4),
                          lower.conf = c(0, 1, 2, 3),
                          upper.conf = c(2, 3, 4, 5),
                          p = c(0.11111, 0.22222, 0.44443, 0.4434343))
prepareDeskStatTable(data = iris, infStat = irisInfStat)
https://cran.r-project.org/web/packages/kableExtra/vignettes/awesome_table_in_html.html
}
\author{
Jochen Kruppa
}
