% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{print_log_reg}
\alias{print_log_reg}
\title{Wrapper function for log reg}
\usage{
print_log_reg(fit, kable_lgl = FALSE)
}
\arguments{
\item{fit}{}

\item{kable_lgl}{}
}
\value{

}
\description{
Wrapper function for log reg
}
\details{
Wrapper function for log reg
}
\author{
Jochen Kruppa
}
