% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sample_tbl}
\alias{Sample_tbl}
\title{Sample metadata}
\format{
A data frame containing sample identifiers, sample types,
injection order, dilution factors, batch information, and statistical
groups.
}
\source{
Example data included with MSlineaR.
}
\usage{
Sample_tbl
}
\description{
Example metadata for the samples contained in the MSlineaR example data.
}
\keyword{datasets}
