% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{write_xlsx}
\alias{write_xlsx}
\title{A wrapper how saves a list to a xls file}
\usage{
write_xlsx(list, file, oneFile = TRUE)
}
\arguments{
\item{list}{The data frames in a list which should be written}

\item{file}{Save location}

\item{oneFile}{One file with sheets or many files?}
}
\description{
A wrapper how saves a list to a xls file
}
\details{
A wrapper how saves a list to a xls file
}
\author{
Jochen Kruppa
}
