% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessLinearity.R
\name{AssessLinearity}
\alias{AssessLinearity}
\title{Ti#' Title}
\usage{
AssessLinearity(
  columnNames,
  dat,
  nCore = 4,
  OutlierResPre = 2,
  minCorFittingModelFOD = 0.99,
  minCorFittingModel = 0.99,
  minConsecutives = 5,
  calcFor = c("Replicates", "Median", "both"),
  Concentration = "Concentration",
  Intensity = "Intensity",
  res = 10,
  listIgnore = NULL,
  levelnoise = 0,
  ...
)
}
\arguments{
\item{minConsecutives}{}
}
\description{
Ti#' Title
}
