% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessLinearity.R
\name{AssessLinearity}
\alias{AssessLinearity}
\title{Title}
\usage{
AssessLinearity(
  COLNAMES = c(ID = "featNames", REPLICATE = NULL, X = "Concentration", Y = "area.raw"),
  DAT,
  nCORE = 1,
  MIN_FEATURE = 5,
  LOG_TRANSFORM = TRUE,
  R2min = 0.9,
  ...
)
}
\arguments{
\item{COLNAMES}{colum names from the input data set to define: ID, number of Replicate, x-value(e.g. Concentration or Dilution) and the y-value(e.g. Intensity or area)}

\item{DAT}{data set with minimum following columns must be provided: ID, REPLICATE,Concentration or Dilution and Intensity or Area}

\item{nCORE}{number of cores to use for parallelization. Default value is 1}

\item{MIN_FEATURE}{minimum number of features within a dilution or concentration series to be considered as linear range}

\item{LOG_TRANSFORM}{logical Value. should the data be log transformed? Default is TRUE.}
}
\description{
Title
}
