% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{checkData}
\alias{checkData}
\title{Prepare raw data}
\usage{
checkData(dat, MIN_FEATURE, ...)
}
\arguments{
\item{...}{}
}
\value{
Data frame with columns:
\itemize{
\item IDintern,
\item ID,
\item IntensityNorm,
\item DP,
\item Comment,
\item pch,
\item color,
\item ConcentrationRaw,
\item IntensityRaw,
\item ConcentrationLog,
\item IntensityLog,
\item IntensityRaw
\item DilutionPoint
}
}
\description{
Prepare raw data
}
