% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{checkData}
\alias{checkData}
\title{Prepare raw data}
\usage{
checkData(
  dat,
  MIN_FEATURE = parent.frame()$MIN_FEATURE,
  TYPE = parent.frame()$TYPE,
  QC = parent.frame()$QC,
  BLANK = parent.frame()$BLANK,
  SAMPLE = parent.frame()$SAMPLE,
  CALIBRANTS = parent.frame()$CALIBRANTS,
  COLNAMES = parent.frame()$COLNAMES,
  NOISE = parent.frame()$NOISE,
  TRANSFORM = parent.frame()$TRANSFORM,
  TRANSFORM_X = parent.frame()$TRANSFORM_X,
  INVERSE_X = parent.frame()$INVERSE_X,
  TRANSFORM_Y = parent.frame()$TRANSFORM_Y,
  INVERSE_Y = parent.frame()$INVERSE_Y,
  FOD = parent.frame()$FOD,
  FOD_MODEL = parent.frame()$FOD_MODEL,
  FOD_SDRES_MIN = parent.frame()$FOD_SDRES_MIN,
  FOD_STDRES_MAX = parent.frame()$FOD_STDRES_MAX,
  TRIMM = parent.frame()$TRIMM,
  SOD = parent.frame()$SOD,
  SOD_MODEL = parent.frame()$SOD_MODEL,
  SOD_SDRES_MIN = parent.frame()$SOD_SDRES_MIN,
  SOD_STDRES_MAX = parent.frame()$SOD_STDRES_MAX,
  LR_SD_RES_FACTOR = parent.frame()$LR_SD_RES_FACTOR,
  R2_MIN = parent.frame()$R2_MIN,
  nCORE = parent.frame()$nCORE,
  GET_OUTPUT = parent.frame()$GET_OUTPUT,
  IMG_OUTPUT_DIR = parent.frame()$IMG_OUTPUT_DIR
)
}
\arguments{
\item{dat}{}
}
\value{
Data frame with columns:
\itemize{
\item IDintern,
\item ID,
\item IntensityNorm,
\item DP,
\item Comment,
\item pch,
\item color,
\item ConcentrationRaw,
\item IntensityRaw,
\item ConcentrationLog,
\item IntensityLog,
\item IntensityRaw
\item DilutionPoint
}
}
\description{
Prepare raw data
}
