% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{checkData}
\alias{checkData}
\title{Prepare raw data}
\usage{
checkData(
  dat,
  MIN_FEATURE,
  TYPE,
  QC,
  QC_REF,
  BLANK,
  SAMPLE,
  SAMPLE_ID,
  CALIBRANTS,
  COLNAMES,
  Y_SAMPLE,
  TRANSFORM,
  TRANSFORM_X,
  INVERSE_X,
  TRANSFORM_Y,
  INVERSE_Y,
  FOD,
  FOD_MODEL,
  FOD_SDRES_MIN,
  FOD_STDRES_MAX,
  TRIMM,
  SOD,
  SOD_MODEL,
  SOD_SDRES_MIN,
  SOD_STDRES_MAX,
  LR_SD_RES_FACTOR,
  R2_MIN,
  BATCH_HARMONIZATION,
  CAL_CONC,
  GET_LR_STATUS,
  nCORE,
  GET_OUTPUT,
  PREFIX,
  OUTPUT_DIR
)
}
\arguments{
\item{dat}{}
}
\value{
Data frame with columns:
\itemize{
\item IDintern,
\item ID,
\item IntensityNorm,
\item DP,
\item Comment,
\item pch,
\item color,
\item ConcentrationRaw,
\item IntensityRaw,
\item ConcentrationLog,
\item IntensityLog,
\item IntensityRaw
\item DilutionPoint
}
}
\description{
Prepare raw data
}
