% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{checkData}
\alias{checkData}
\title{Prepare raw data}
\usage{
checkData(DAT, ...)
}
\arguments{
\item{dat}{Data frame with raw data. Mandatory columns are:
\itemize{
\item ID,
\item IDintern
\item Batch,
\item Replicate,
\item MZ
\item RT
\item Concentration
\item Intensity
}}
}
\value{
Data frame with columns:
\itemize{
\item IDintern,
\item ID,
\item IntensityNorm,
\item DP,
\item Comment,
\item pch,
\item color,
\item ConcentrationRaw,
\item IntensityRaw,
\item ConcentrationLog,
\item IntensityLog,
\item IntensityRaw
\item DilutionPoint
}
}
\description{
Prepare raw data
}
