% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittingModel.R
\name{chooseModel}
\alias{chooseModel}
\title{Title}
\usage{
chooseModel(
  dats,
  y = parent.frame()$Y,
  x = parent.frame()$X,
  model = c("logistic", "linear", "quadratic"),
  SDRES_MIN = 0,
  STDRES = 2,
  abbr
)
}
\arguments{
\item{model}{}
}
\description{
Title
}
