% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittingModel.R
\name{chooseModel}
\alias{chooseModel}
\title{Title}
\usage{
chooseModel(
  dat,
  y = "IntensityNorm",
  x = "DP",
  model = c("logistic", "linear", "quadratic")
)
}
\arguments{
\item{model}{}
}
\description{
Title
}
