% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{combineData}
\alias{combineData}
\title{Title}
\usage{
combineData(
  inputData_Series,
  inputData_BioSamples,
  inputData_QC,
  inputData_QC_ref,
  inputData_Blank,
  columns = c(ID = COLNAMES[["ID"]], Batch = COLNAMES[["Batch"]], X = Xraw, Y = Yraw),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
