% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMinimumValues.R
\name{countMinimumValue}
\alias{countMinimumValue}
\title{Title}
\usage{
countMinimumValue(DAT, MIN_FEATURE = 3, step)
}
\arguments{
\item{MIN_FEATURE}{minimum number of features within a dilution or concentration series to be considered as linear range}

\item{Y}{}
}
\description{
Title
}
