% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLinearRange.R
\name{create_output_findLinearRange}
\alias{create_output_findLinearRange}
\title{calculate linear regression from choosen range and generate output table for function \code{findLinearRange()}}
\usage{
create_output_findLinearRange(
  inRange,
  data,
  y = y,
  x = x,
  real_x = real_x,
  min_feature = min_feature
)
}
\arguments{
\item{inRange}{Boolean; Does the feature has a choosen range?}

\item{data}{Long format data frame or data table for one metabolite generated by function \code{findLinearRange()}}

\item{y}{String; Column name of the log transformed dependent variable}

\item{x}{String; Column name of the log transformed independent variable}
}
\value{
list including two tables with metrics for each signal in the first one and information about the whole feature in the second one
}
\description{
calculate linear regression from choosen range and generate output table for function \code{findLinearRange()}
}
