% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLinearRange.R
\name{findLinearRange}
\alias{findLinearRange}
\title{Title}
\usage{
findLinearRange(
  dats,
  x = "DilutionPoint",
  y = "IntensityNorm",
  modelObject,
  res = 20,
  MIN_FEATURE
)
}
\arguments{
\item{res}{}
}
\description{
Title
}
