% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLinearRange.R
\name{findLinearRange}
\alias{findLinearRange}
\title{Title}
\usage{
findLinearRange(
  dats,
  x = "DilutionPoint",
  y = "IntensityNorm",
  sd_res_factor = 2,
  min_feature,
  real_x
)
}
\arguments{
\item{res}{}
}
\description{
Title
}
