% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimEnds.R
\name{findPlateaus}
\alias{findPlateaus}
\title{Title}
\usage{
findPlateaus(dats, y = y, x = x, slope_ratio = 0.5)
}
\arguments{
\item{dats}{Long format data frame or data table for one metabolite including
information about dependent and independent variable. Called from function \code{TrimmEnds()}}

\item{y}{String; Column name of the log transformed dependent variable}

\item{x}{String; Column name of the log transformed independent variable}

\item{slope_ratio}{Integer; factor, describing the allwid difference between
the calculated slope and the middle slope of the curve. Default 0.5, which means
if the slopes at the beginning or the end is equal or less than half of the
slope in the middle, the dilution points are probably in a plateau and will be removed.}
}
\value{
vector of booleans
}
\description{
Title
}
