% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSamples.R
\name{getConc}
\alias{getConc}
\title{Title}
\usage{
getConc(
  dats,
  datCal,
  COLNAMES = c(ID = "Compound", Replicate = "Batch", Y = "Area")
)
}
\arguments{
\item{dats}{samples compare to the dilution series}

\item{datCal}{\link{output object from AssessLinearity function}$summaryFDS}

\item{COLNAMES}{}
}
\description{
Title
}
