% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSamples.R
\name{getConc}
\alias{getConc}
\title{Title}
\usage{
getConc(dats, datCal, y, INVERSE_Y, NAME_Standard, COL_expConc)
}
\arguments{
\item{dats}{data table with information about the biological samples.
Required columns are IDintern, y, groupIndices.It is necessary
that the columns in dats and datCal have the same column names and that the
column groupIndices is compatible between samples and dilution/concentration curve data.}

\item{datCal}{data table with dilution/concentration curve information of the linearity.
Required columns are groupIndices, Intercept and slope. It is necessary
that the columns in dats and datCal have the same column names and that the
column groupIndices is compatible between samples and dilution/concentration curve data.}

\item{INVERSE_Y}{}
}
\description{
Title
}
