% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSamples.R
\name{getLRstatus}
\alias{getLRstatus}
\title{Title}
\usage{
getLRstatus(
  LR_object,
  dats,
  COLNAMES = c(ID = "Compound", Replicate = "Batch", Y = "Area")
)
}
\arguments{
\item{dats}{samples compare to the dilution series}

\item{COLNAMES}{}

\item{datCal}{output object from AssessLinearity}
}
\description{
Title
}
