% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterData.R
\name{MS_filterSamples}
\alias{MS_filterSamples}
\title{Title}
\usage{
MS_filterSamples(
  inputData,
  Compound_ID,
  Sample_ID,
  Y,
  inputDataSummary,
  outputformat = c("wide", "long", "both"),
  PATH,
  filter.type = c("Batch", "Class", "Batch_Class"),
  filter.calc = c("all", "at least one", "at least x"),
  filter.calc.x.Batch = NULL,
  filter.calc.x.Class = NULL,
  filter.percent = 80
)
}
\arguments{
\item{filter.percent}{}
}
\description{
Title
}
