% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_Barplot_Summary}
\alias{plot_Barplot_Summary}
\title{Title}
\usage{
plot_Barplot_Summary(
  inputData_Series,
  printPDF = TRUE,
  GroupIndices = "all",
  Feature = "all",
  outputfileName = c("Summary_Barplot"),
  columns = c(ID = COLNAMES[["ID"]], Batch = COLNAMES[["Batch"]], X = Xraw, Y = Yraw),
  ...
)
}
\arguments{
\item{pdfheight}{}
}
\description{
Title
}
