% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_Barplot_Summary}
\alias{plot_Barplot_Summary}
\title{Title}
\usage{
plot_Barplot_Summary(
  inputData_Series,
  printPDF = TRUE,
  GroupIndices = "all",
  Feature = "all",
  outputfileName = c("Summary_Barplot"),
  COLNAMES,
  X,
  Y,
  output_dir
)
}
\arguments{
\item{outputfileName}{}
}
\description{
Title
}
