% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_Barplot_Summary_Sample}
\alias{plot_Barplot_Summary_Sample}
\title{Title}
\usage{
plot_Barplot_Summary_Sample(
  inputData_Samples,
  printPDF = TRUE,
  GroupIndices = "all",
  Feature = "all",
  outputfileName = c("Summary_Barplot_Samples"),
  columns = c(ID = COLNAMES[["ID"]], Batch = COLNAMES[["Batch"]], X = Xraw, Y = Yraw),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
