% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_FDS}
\alias{plot_FDS}
\title{Title}
\usage{
plot_FDS(
  inputData_Series,
  inputData_BioSamples,
  inputData_QC,
  inputData_QC_ref = NULL,
  inputData_Blank = NULL,
  nrRow = 10,
  nrFeature = 50,
  printPDF = TRUE,
  GroupIndices = "all",
  Feature = "all",
  printR2 = TRUE,
  outputfileName = c("Calibrationplot"),
  columns = c(ID = COLNAMES[["ID"]], Batch = COLNAMES[["Batch"]], X = Xraw, Y = Yraw),
  ...
)
}
\arguments{
\item{outputfileName}{}
}
\description{
Title
}
