% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_FDS}
\alias{plot_FDS}
\title{Title}
\usage{
plot_FDS(
  inputData_Series,
  inputData_BioSamples,
  inputData_QC,
  nrFeature = 50,
  printPDF = TRUE,
  GroupIndices = "all",
  Feature = "all",
  printR2 = TRUE,
  outputfileName = c("Calibrationplot"),
  TRANSFORM_Y,
  inverse_y,
  COLNAMES,
  X,
  Y,
  Series,
  output_dir
)
}
\arguments{
\item{outputfileName}{}
}
\description{
Title
}
