% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prepareData}
\alias{prepareData}
\title{Title
normalizing, centralizing, log transforming}
\usage{
prepareData(
  dat,
  TRANSFORM = parent.frame()$TRANSFORM,
  TRANSFORM_X = parent.frame()$TRANSFORM_X,
  TRANSFORM_Y = parent.frame()$TRANSFORM_Y,
  COLNAMES = parent.frame()$COLNAMES,
  TYPE = parent.frame()$TYPE
)
}
\arguments{
\item{dat}{}
}
\description{
Title
normalizing, centralizing, log transforming
}
