% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prepareData}
\alias{prepareData}
\title{Title
normalizing, centralizing, log transforming}
\usage{
prepareData(dat, ...)
}
\arguments{
\item{dat}{}
}
\description{
Title
normalizing, centralizing, log transforming
}
